Model Test 분리.
